requirements.txt
Flask
tensorflow
keras
Pillow
numpy
joblib
opencv-python
